server/requirements.txt
simple-websocket-server==0.4.4
websockets==11.0.3
python-dotenv==1.0.1
